+++ src/lib.rs
@@ -6,9 +6,7 @@ use language_servers::LanguageServer;
 use zed_extension_api::LanguageServerId;
 use zed_extension_api::Result;
 use zed_extension_api::Worktree;
-use zed_extension_api::{
-    self as zed,
-};
+use zed_extension_api::{self as zed};
 
 #[derive(Default)]
 struct DjangoExtension {
@@ -40,6 +38,33 @@ impl zed::Extension for DjangoExtension {
                 server.language_server_command(language_server_id, worktree)
             }
             language_server_id => Err(format!("unknown language server: {language_server_id}")),
+        }
+    }
+
+    fn language_server_initialization_options(
+        &mut self,
+        language_server_id: &LanguageServerId,
+        worktree: &Worktree,
+    ) -> Result<Option<zed::serde_json::Value>> {
+        match language_server_id.as_ref() {
+            DjangoLanguageServer::SERVER_ID => {
+                let server = self
+                    .django_language_server
+                    .get_or_insert_with(DjangoLanguageServer::new);
+                server.language_server_initialization_options(language_server_id, worktree)
+            }
+            DjangoTemplateLsp::SERVER_ID => {
+                let server = self
+                    .django_template_lsp
+                    .get_or_insert_with(DjangoTemplateLsp::new);
+                server.language_server_initialization_options(language_server_id, worktree)
+            }
+            _ => Ok(Some(
+                zed::settings::LspSettings::for_worktree(language_server_id.as_ref(), worktree)
+                    .ok()
+                    .and_then(|lsp_settings| lsp_settings.initialization_options.clone())
+                    .unwrap_or_default(),
+            )),
         }
     }
 }